reference_for_heatmap/ai_detector/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pillow==10.4.0
transformers==4.44.0
torch==2.4.0
torchvision==0.19.0
python-multipart==0.0.9
opencv-python==4.10.0.84
numpy==1.26.4
